Fix #1. Reported by @erikarvstedt. check_fuser() fails.
The curent test for check_fuser() is
return (WIFEXITED(ret) && WEXITSTATUS(ret) == 0);

Now, a separate test is made for WIFEXITED because even if a file is in
use, it can still return a SUCCESS status. In addition, WEXITSTATUS
should only be tested if WIFEXITED is true. And, paradoxically,
WEXITSTATUS returns zero if there is no error.

check_fuser will return an error if WIFEXITED is zero (unlikely) or if
WIFEXITED is true and WEXITSTATUS is non-zero.